+++ lib/eventum/class.access.php
@@ -645,8 +645,8 @@ class Access
                                 SUBSTR(iss_access_level, 1, 6) = 'group_' AND ugr_grp_id = SUBSTR(iss_access_level, 7)
                             ) OR
                             (
-                                isu_usr_id " . Auth::getUserID() . "
-                            )";
+                                isu_usr_id = " . Auth::getUserID() . '
+                            )';
 
             $workflow = Workflow::getAdditionalAccessSQL($prj_id, Auth::getUserID());
             if ($workflow != null) {